+++ src/frequenz/client/dispatch/_client.py
@@ -7,8 +7,11 @@ from __future__ import annotations
 from datetime import datetime, timedelta
 from typing import Any, AsyncIterator, Awaitable, Iterator, Literal, cast
 
-# pylint: disable=no-name-in-module
-from frequenz.api.common.v1.pagination.pagination_params_pb2 import PaginationParams
+# pylint: disable-next=no-name-in-module
+from frequenz.api.common.v1alpha8.pagination.pagination_params_pb2 import (
+    PaginationParams,
+)
+from frequenz.api.common.v1alpha8.types.interval_pb2 import Interval as PBInterval
 from frequenz.api.dispatch.v1 import dispatch_pb2_grpc
 from frequenz.api.dispatch.v1.dispatch_pb2 import (
     CreateMicrogridDispatchResponse,
@@ -20,11 +23,6 @@ from frequenz.api.dispatch.v1.dispatch_pb2 import (
     ListMicrogridDispatchesResponse,
     StreamMicrogridDispatchesRequest,
     StreamMicrogridDispatchesResponse,
-)
-from frequenz.api.dispatch.v1.dispatch_pb2 import (
-    TimeIntervalFilter as PBTimeIntervalFilter,
-)
-from frequenz.api.dispatch.v1.dispatch_pb2 import (
     UpdateMicrogridDispatchRequest,
     UpdateMicrogridDispatchResponse,
 )
@@ -170,11 +168,9 @@ class DispatchApiClient(BaseApiClient[dispatch_pb2_grpc.MicrogridDispatchService
 
         def to_interval(
             from_: datetime | None, to: datetime | None
-        ) -> PBTimeIntervalFilter | None:
+        ) -> PBInterval | None:
             return (
-                PBTimeIntervalFilter(
-                    from_time=to_timestamp(from_), to_time=to_timestamp(to)
-                )
+                PBInterval(start_time=to_timestamp(from_), end_time=to_timestamp(to))
                 if from_ or to
                 else None
             )